significant changes to change foreign key references to the lookup data (countries, carriers, airports) to strings.  updated corresponding UDR mapping files, but will likely require updating front-end code.

+++ gtas-parent/gtas-commons/src/main/java/gov/gtas/model/Itinerary.java
@@ -12,7 +12,7 @@ public class Itinerary extends BaseEntity {
     @JoinColumn
     private Flight flight;
     
-    private Short legNumber;
+    private Integer legNumber;
     
     @ManyToOne
     @JoinColumn
